Move --dest test to cmd-dest.test.js like in Bubbleprof
