+++ visualizer/d3.js
@@ -7,7 +7,6 @@ const d3 = Object.assign(
   selection
 )
 
-
 // This property changes after importing so we fake a live binding.
 Object.defineProperty(d3, 'event', {
   get () { return selection.event }
